This is a first attempt at getting the Ruby Run command working on 1.9 (while keeping 1.8 support).  I doubt I have located all of the edge cases yet, but it seems to be running simple scripts for me now, with and without errors in them.
I'm not sure if this is the best way to handle encoding, since I'm forcing a -KU on the executed code in both 1.8 and 1.9 now.  This should just change the default source encoding for 1.9 and hopefully not break code since that's user overridable.  It's probably more hostile in Ruby 1.8, but our executor library was already doing this in a different way.  If you don't like this policy, email me suggestions.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Ruby.tmbundle@11067 dfb7d73b-c2ec-0310-8fea-fb051d288c6d